chore(*): refactoring the splitter for partial ivy
