+++ projects/igniteui-angular/src/lib/splitter/splitter-bar.component.html
