+++ projects/igniteui-angular/src/lib/splitter/splitter.module.ts
@@ -1,8 +1,7 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { IgxSplitBarComponent } from './splitter-bar/splitter-bar.component';
 import { IgxSplitterPaneComponent } from './splitter-pane/splitter-pane.component';
-import { IgxSplitterComponent } from './splitter.component';
+import { IgxSplitterComponent, IgxSplitBarComponent } from './splitter.component';
 import { IgxIconModule } from '../icon/public_api';
 import { IgxDragDropModule } from '../directives/drag-drop/drag-drop.directive';
